# go-beego-exapmple-04-todo/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "beego Test API",
        "description": "beego has a very cool tools to autogenerate documents for your API",
        "version": "1.0.0",
        "termsOfService": "http://beego.me/",
        "contact": {
            "email": "astaxie@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
        "/todo/": {
            "get": {
                "tags": [
                    "todo"
                ],
                "description": "get Todo",
                "operationId": "TodoController.Get All",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    },
                    "403": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "todo"
                ],
                "description": "create Todo",
                "operationId": "TodoController.Post",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for Todo content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Todo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{int} models.Todo"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Todo": {
            "title": "Todo",
            "type": "object",
            "properties": {
                "Completed": {
                    "type": "boolean"
                },
                "Id": {
                    "type": "integer",
                    "format": "int64"
                },
                "Title": {
                    "type": "string"
                }
            }
        }
    },
...
}